sui_rpc

Module headers

Constants

X_SUI_CHAIN

Chain name of the current chain

X_SUI_CHAIN_ID

Chain ID of the current chain

X_SUI_CHECKPOINT_HEIGHT

Current checkpoint height

X_SUI_EPOCH

Current epoch of the chain

X_SUI_LOWEST_AVAILABLE_CHECKPOINT

Lowest available checkpoint for which transaction and checkpoint data can be requested.

X_SUI_LOWEST_AVAILABLE_CHECKPOINT_OBJECTS

Lowest available checkpoint for which object data can be requested.

X_SUI_TIMESTAMP

Current timestamp of the chain - encoded in the RFC 3339 format.

X_SUI_TIMESTAMP_MS

Current timestamp of the chain - represented as number of milliseconds from the Unix epoch