sui_rpc::proto::sui::rpc::v2

Struct ChangedObject

#[non_exhaustive]
pub struct ChangedObject {
    pub object_id: Option<String>,
    pub input_state: Option<i32>,
    pub input_version: Option<u64>,
    pub input_digest: Option<String>,
    pub input_owner: Option<Owner>,
    pub output_state: Option<i32>,
    pub output_version: Option<u64>,
    pub output_digest: Option<String>,
    pub output_owner: Option<Owner>,
    pub accumulator_write: Option<AccumulatorWrite>,
    pub id_operation: Option<i32>,
    pub object_type: Option<String>,
}

Input/output state of an object that was changed during execution.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

object_id: Option<String>

ID of the object.

input_state: Option<i32>input_version: Option<u64>

Version of the object before this transaction executed.

input_digest: Option<String>

Digest of the object before this transaction executed.

input_owner: Option<Owner>

Owner of the object before this transaction executed.

output_state: Option<i32>output_version: Option<u64>

Version of the object after this transaction executed.

output_digest: Option<String>

Digest of the object after this transaction executed.

output_owner: Option<Owner>

Owner of the object after this transaction executed.

accumulator_write: Option<AccumulatorWrite>

The contents of the accumulator write when output_state is OUTPUT_OBJECT_STATE_ACCUMULATOR_WRITE

id_operation: Option<i32>

What happened to an ObjectId during execution.

object_type: Option<String>

Type information is not provided by the effects structure but is instead provided by an indexing layer

Implementations

impl ChangedObject

pub const OBJECT_ID_FIELD: &'static MessageField

pub const INPUT_STATE_FIELD: &'static MessageField

pub const INPUT_VERSION_FIELD: &'static MessageField

pub const INPUT_DIGEST_FIELD: &'static MessageField

pub const INPUT_OWNER_FIELD: &'static MessageField

pub const OUTPUT_STATE_FIELD: &'static MessageField

pub const OUTPUT_VERSION_FIELD: &'static MessageField

pub const OUTPUT_DIGEST_FIELD: &'static MessageField

pub const OUTPUT_OWNER_FIELD: &'static MessageField

pub const ACCUMULATOR_WRITE_FIELD: &'static MessageField

pub const ID_OPERATION_FIELD: &'static MessageField

pub const OBJECT_TYPE_FIELD: &'static MessageField

impl ChangedObject

pub fn path_builder() -> ChangedObjectFieldPathBuilder

impl ChangedObject

pub const fn const_default() -> Self

pub fn object_id_opt_mut(&mut self) -> Option<&mut String>

If object_id is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn object_id_mut(&mut self) -> &mut String

Returns a mutable reference to object_id. If the field is unset, it is first initialized with the default value.

pub fn object_id_opt(&self) -> Option<&str>

If object_id is set, returns Some with the value; otherwise returns None.

pub fn set_object_id<T: Into<String>>(&mut self, field: T)

Sets object_id with the provided value.

pub fn with_object_id<T: Into<String>>(self, field: T) -> Self

Sets object_id with the provided value.

pub fn with_input_state<T: Into<InputObjectState>>(self, field: T) -> Self

Sets input_state with the provided value.

pub fn input_version_opt_mut(&mut self) -> Option<&mut u64>

If input_version is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn input_version_mut(&mut self) -> &mut u64

Returns a mutable reference to input_version. If the field is unset, it is first initialized with the default value.

pub fn input_version_opt(&self) -> Option<u64>

If input_version is set, returns Some with the value; otherwise returns None.

pub fn set_input_version(&mut self, field: u64)

Sets input_version with the provided value.

pub fn with_input_version(self, field: u64) -> Self

Sets input_version with the provided value.

pub fn input_digest_opt_mut(&mut self) -> Option<&mut String>

If input_digest is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn input_digest_mut(&mut self) -> &mut String

Returns a mutable reference to input_digest. If the field is unset, it is first initialized with the default value.

pub fn input_digest_opt(&self) -> Option<&str>

If input_digest is set, returns Some with the value; otherwise returns None.

pub fn set_input_digest<T: Into<String>>(&mut self, field: T)

Sets input_digest with the provided value.

pub fn with_input_digest<T: Into<String>>(self, field: T) -> Self

Sets input_digest with the provided value.

pub fn input_owner(&self) -> &Owner

Returns the value of input_owner, or the default value if input_owner is unset.

pub fn input_owner_opt_mut(&mut self) -> Option<&mut Owner>

If input_owner is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn input_owner_mut(&mut self) -> &mut Owner

Returns a mutable reference to input_owner. If the field is unset, it is first initialized with the default value.

pub fn input_owner_opt(&self) -> Option<&Owner>

If input_owner is set, returns Some with the value; otherwise returns None.

pub fn set_input_owner<T: Into<Owner>>(&mut self, field: T)

Sets input_owner with the provided value.

pub fn with_input_owner<T: Into<Owner>>(self, field: T) -> Self

Sets input_owner with the provided value.

pub fn with_output_state<T: Into<OutputObjectState>>(self, field: T) -> Self

Sets output_state with the provided value.

pub fn output_version_opt_mut(&mut self) -> Option<&mut u64>

If output_version is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn output_version_mut(&mut self) -> &mut u64

Returns a mutable reference to output_version. If the field is unset, it is first initialized with the default value.

pub fn output_version_opt(&self) -> Option<u64>

If output_version is set, returns Some with the value; otherwise returns None.

pub fn set_output_version(&mut self, field: u64)

Sets output_version with the provided value.

pub fn with_output_version(self, field: u64) -> Self

Sets output_version with the provided value.

pub fn output_digest_opt_mut(&mut self) -> Option<&mut String>

If output_digest is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn output_digest_mut(&mut self) -> &mut String

Returns a mutable reference to output_digest. If the field is unset, it is first initialized with the default value.

pub fn output_digest_opt(&self) -> Option<&str>

If output_digest is set, returns Some with the value; otherwise returns None.

pub fn set_output_digest<T: Into<String>>(&mut self, field: T)

Sets output_digest with the provided value.

pub fn with_output_digest<T: Into<String>>(self, field: T) -> Self

Sets output_digest with the provided value.

pub fn output_owner(&self) -> &Owner

Returns the value of output_owner, or the default value if output_owner is unset.

pub fn output_owner_opt_mut(&mut self) -> Option<&mut Owner>

If output_owner is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn output_owner_mut(&mut self) -> &mut Owner

Returns a mutable reference to output_owner. If the field is unset, it is first initialized with the default value.

pub fn output_owner_opt(&self) -> Option<&Owner>

If output_owner is set, returns Some with the value; otherwise returns None.

pub fn set_output_owner<T: Into<Owner>>(&mut self, field: T)

Sets output_owner with the provided value.

pub fn with_output_owner<T: Into<Owner>>(self, field: T) -> Self

Sets output_owner with the provided value.

pub fn accumulator_write(&self) -> &AccumulatorWrite

Returns the value of accumulator_write, or the default value if accumulator_write is unset.

pub fn accumulator_write_opt_mut(&mut self) -> Option<&mut AccumulatorWrite>

If accumulator_write is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn accumulator_write_mut(&mut self) -> &mut AccumulatorWrite

Returns a mutable reference to accumulator_write. If the field is unset, it is first initialized with the default value.

pub fn accumulator_write_opt(&self) -> Option<&AccumulatorWrite>

If accumulator_write is set, returns Some with the value; otherwise returns None.

pub fn set_accumulator_write<T: Into<AccumulatorWrite>>(&mut self, field: T)

Sets accumulator_write with the provided value.

pub fn with_accumulator_write<T: Into<AccumulatorWrite>>(self, field: T) -> Self

Sets accumulator_write with the provided value.

pub fn with_id_operation<T: Into<IdOperation>>(self, field: T) -> Self

Sets id_operation with the provided value.

pub fn object_type_opt_mut(&mut self) -> Option<&mut String>

If object_type is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn object_type_mut(&mut self) -> &mut String

Returns a mutable reference to object_type. If the field is unset, it is first initialized with the default value.

pub fn object_type_opt(&self) -> Option<&str>

If object_type is set, returns Some with the value; otherwise returns None.

pub fn set_object_type<T: Into<String>>(&mut self, field: T)

Sets object_type with the provided value.

pub fn with_object_type<T: Into<String>>(self, field: T) -> Self

Sets object_type with the provided value.

impl ChangedObject

pub fn object_id(&self) -> &str

Returns the value of object_id, or the default value if object_id is unset.

pub fn input_state(&self) -> InputObjectState

Returns the enum value of input_state, or the default if the field is unset or set to an invalid enum value.

pub fn set_input_state(&mut self, value: InputObjectState)

Sets input_state to the provided enum value.

pub fn input_version(&self) -> u64

Returns the value of input_version, or the default value if input_version is unset.

pub fn input_digest(&self) -> &str

Returns the value of input_digest, or the default value if input_digest is unset.

pub fn output_state(&self) -> OutputObjectState

Returns the enum value of output_state, or the default if the field is unset or set to an invalid enum value.

pub fn set_output_state(&mut self, value: OutputObjectState)

Sets output_state to the provided enum value.

pub fn output_version(&self) -> u64

Returns the value of output_version, or the default value if output_version is unset.

pub fn output_digest(&self) -> &str

Returns the value of output_digest, or the default value if output_digest is unset.

pub fn id_operation(&self) -> IdOperation

Returns the enum value of id_operation, or the default if the field is unset or set to an invalid enum value.

pub fn set_id_operation(&mut self, value: IdOperation)

Sets id_operation to the provided enum value.

pub fn object_type(&self) -> &str

Returns the value of object_type, or the default value if object_type is unset.