sui_rpc::proto::sui::rpc::v2

Struct Event

#[non_exhaustive]
pub struct Event {
    pub package_id: Option<String>,
    pub module: Option<String>,
    pub sender: Option<String>,
    pub event_type: Option<String>,
    pub contents: Option<Bcs>,
    pub json: Option<Box<Value>>,
}

An event.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

package_id: Option<String>

Package ID of the top-level function invoked by a MoveCall command that triggered this event to be emitted.

module: Option<String>

Module name of the top-level function invoked by a MoveCall command that triggered this event to be emitted.

sender: Option<String>

Address of the account that sent the transaction where this event was emitted.

event_type: Option<String>

The type of the event emitted.

contents: Option<Bcs>

BCS serialized bytes of the event.

json: Option<Box<Value>>

JSON rendering of the event.

Implementations

impl Event

pub const PACKAGE_ID_FIELD: &'static MessageField

pub const MODULE_FIELD: &'static MessageField

pub const SENDER_FIELD: &'static MessageField

pub const EVENT_TYPE_FIELD: &'static MessageField

pub const CONTENTS_FIELD: &'static MessageField

pub const JSON_FIELD: &'static MessageField

impl Event

pub fn path_builder() -> EventFieldPathBuilder

impl Event

pub const fn const_default() -> Self

pub fn package_id_opt_mut(&mut self) -> Option<&mut String>

If package_id is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn package_id_mut(&mut self) -> &mut String

Returns a mutable reference to package_id. If the field is unset, it is first initialized with the default value.

pub fn package_id_opt(&self) -> Option<&str>

If package_id is set, returns Some with the value; otherwise returns None.

pub fn set_package_id<T: Into<String>>(&mut self, field: T)

Sets package_id with the provided value.

pub fn with_package_id<T: Into<String>>(self, field: T) -> Self

Sets package_id with the provided value.

pub fn module_opt_mut(&mut self) -> Option<&mut String>

If module is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn module_mut(&mut self) -> &mut String

Returns a mutable reference to module. If the field is unset, it is first initialized with the default value.

pub fn module_opt(&self) -> Option<&str>

If module is set, returns Some with the value; otherwise returns None.

pub fn set_module<T: Into<String>>(&mut self, field: T)

Sets module with the provided value.

pub fn with_module<T: Into<String>>(self, field: T) -> Self

Sets module with the provided value.

pub fn sender_opt_mut(&mut self) -> Option<&mut String>

If sender is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn sender_mut(&mut self) -> &mut String

Returns a mutable reference to sender. If the field is unset, it is first initialized with the default value.

pub fn sender_opt(&self) -> Option<&str>

If sender is set, returns Some with the value; otherwise returns None.

pub fn set_sender<T: Into<String>>(&mut self, field: T)

Sets sender with the provided value.

pub fn with_sender<T: Into<String>>(self, field: T) -> Self

Sets sender with the provided value.

pub fn event_type_opt_mut(&mut self) -> Option<&mut String>

If event_type is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn event_type_mut(&mut self) -> &mut String

Returns a mutable reference to event_type. If the field is unset, it is first initialized with the default value.

pub fn event_type_opt(&self) -> Option<&str>

If event_type is set, returns Some with the value; otherwise returns None.

pub fn set_event_type<T: Into<String>>(&mut self, field: T)

Sets event_type with the provided value.

pub fn with_event_type<T: Into<String>>(self, field: T) -> Self

Sets event_type with the provided value.

pub fn contents(&self) -> &Bcs

Returns the value of contents, or the default value if contents is unset.

pub fn contents_opt_mut(&mut self) -> Option<&mut Bcs>

If contents is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn contents_mut(&mut self) -> &mut Bcs

Returns a mutable reference to contents. If the field is unset, it is first initialized with the default value.

pub fn contents_opt(&self) -> Option<&Bcs>

If contents is set, returns Some with the value; otherwise returns None.

pub fn set_contents<T: Into<Bcs>>(&mut self, field: T)

Sets contents with the provided value.

pub fn with_contents<T: Into<Bcs>>(self, field: T) -> Self

Sets contents with the provided value.

pub fn json_opt_mut(&mut self) -> Option<&mut Value>

If json is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn json_mut(&mut self) -> &mut Value

Returns a mutable reference to json. If the field is unset, it is first initialized with the default value.

pub fn json_opt(&self) -> Option<&Value>

If json is set, returns Some with the value; otherwise returns None.

pub fn set_json<T: Into<Value>>(&mut self, field: T)

Sets json with the provided value.

pub fn with_json<T: Into<Value>>(self, field: T) -> Self

Sets json with the provided value.

impl Event

pub fn package_id(&self) -> &str

Returns the value of package_id, or the default value if package_id is unset.

pub fn module(&self) -> &str

Returns the value of module, or the default value if module is unset.

pub fn sender(&self) -> &str

Returns the value of sender, or the default value if sender is unset.

pub fn event_type(&self) -> &str

Returns the value of event_type, or the default value if event_type is unset.