sui_rpc::proto::sui::rpc::v2

Struct GetCoinInfoResponse

#[non_exhaustive]
pub struct GetCoinInfoResponse {
    pub coin_type: Option<String>,
    pub metadata: Option<CoinMetadata>,
    pub treasury: Option<CoinTreasury>,
    pub regulated_metadata: Option<RegulatedCoinMetadata>,
}

Response message for NodeService.GetCoinInfo.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

coin_type: Option<String>

Required. The coin type.

metadata: Option<CoinMetadata>

This field will be populated with information about this coin type’s 0x2::coin::CoinMetadata if it exists and has not been wrapped.

treasury: Option<CoinTreasury>

This field will be populated with information about this coin type’s 0x2::coin::TreasuryCap if it exists and has not been wrapped.

regulated_metadata: Option<RegulatedCoinMetadata>

If this coin type is a regulated coin, this field will be populated with information either from its Currency object in the CoinRegistry, or from its 0x2::coin::RegulatedCoinMetadata object for coins that have not been migrated to the CoinRegistry

If this coin is not known to be regulated, only the coin_regulated_state field will be populated.

Implementations

impl GetCoinInfoResponse

pub const COIN_TYPE_FIELD: &'static MessageField

pub const METADATA_FIELD: &'static MessageField

pub const TREASURY_FIELD: &'static MessageField

pub const REGULATED_METADATA_FIELD: &'static MessageField

impl GetCoinInfoResponse

pub fn path_builder() -> GetCoinInfoResponseFieldPathBuilder

impl GetCoinInfoResponse

pub const fn const_default() -> Self

pub fn coin_type_opt_mut(&mut self) -> Option<&mut String>

If coin_type is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn coin_type_mut(&mut self) -> &mut String

Returns a mutable reference to coin_type. If the field is unset, it is first initialized with the default value.

pub fn coin_type_opt(&self) -> Option<&str>

If coin_type is set, returns Some with the value; otherwise returns None.

pub fn set_coin_type<T: Into<String>>(&mut self, field: T)

Sets coin_type with the provided value.

pub fn with_coin_type<T: Into<String>>(self, field: T) -> Self

Sets coin_type with the provided value.

pub fn metadata(&self) -> &CoinMetadata

Returns the value of metadata, or the default value if metadata is unset.

pub fn metadata_opt_mut(&mut self) -> Option<&mut CoinMetadata>

If metadata is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn metadata_mut(&mut self) -> &mut CoinMetadata

Returns a mutable reference to metadata. If the field is unset, it is first initialized with the default value.

pub fn metadata_opt(&self) -> Option<&CoinMetadata>

If metadata is set, returns Some with the value; otherwise returns None.

pub fn set_metadata<T: Into<CoinMetadata>>(&mut self, field: T)

Sets metadata with the provided value.

pub fn with_metadata<T: Into<CoinMetadata>>(self, field: T) -> Self

Sets metadata with the provided value.

pub fn treasury(&self) -> &CoinTreasury

Returns the value of treasury, or the default value if treasury is unset.

pub fn treasury_opt_mut(&mut self) -> Option<&mut CoinTreasury>

If treasury is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn treasury_mut(&mut self) -> &mut CoinTreasury

Returns a mutable reference to treasury. If the field is unset, it is first initialized with the default value.

pub fn treasury_opt(&self) -> Option<&CoinTreasury>

If treasury is set, returns Some with the value; otherwise returns None.

pub fn set_treasury<T: Into<CoinTreasury>>(&mut self, field: T)

Sets treasury with the provided value.

pub fn with_treasury<T: Into<CoinTreasury>>(self, field: T) -> Self

Sets treasury with the provided value.

pub fn regulated_metadata(&self) -> &RegulatedCoinMetadata

Returns the value of regulated_metadata, or the default value if regulated_metadata is unset.

pub fn regulated_metadata_opt_mut( &mut self, ) -> Option<&mut RegulatedCoinMetadata>

If regulated_metadata is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn regulated_metadata_mut(&mut self) -> &mut RegulatedCoinMetadata

Returns a mutable reference to regulated_metadata. If the field is unset, it is first initialized with the default value.

pub fn regulated_metadata_opt(&self) -> Option<&RegulatedCoinMetadata>

If regulated_metadata is set, returns Some with the value; otherwise returns None.

pub fn set_regulated_metadata<T: Into<RegulatedCoinMetadata>>( &mut self, field: T, )

Sets regulated_metadata with the provided value.

pub fn with_regulated_metadata<T: Into<RegulatedCoinMetadata>>( self, field: T, ) -> Self

Sets regulated_metadata with the provided value.

impl GetCoinInfoResponse

pub fn coin_type(&self) -> &str

Returns the value of coin_type, or the default value if coin_type is unset.