sui_rpc::proto::sui::rpc::v2

Struct MoveCall

#[non_exhaustive]
pub struct MoveCall {
    pub package: Option<String>,
    pub module: Option<String>,
    pub function: Option<String>,
    pub type_arguments: Vec<String>,
    pub arguments: Vec<Argument>,
}

Command to call a Move function.

Functions that can be called by a MoveCall command are those that have a function signature that is either entry or public (which don’t have a reference return type).

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

package: Option<String>

The package containing the module and function.

module: Option<String>

The specific module in the package containing the function.

function: Option<String>

The function to be called.

type_arguments: Vec<String>

The type arguments to the function.

arguments: Vec<Argument>

The arguments to the function.

Implementations

impl MoveCall

pub const PACKAGE_FIELD: &'static MessageField

pub const MODULE_FIELD: &'static MessageField

pub const FUNCTION_FIELD: &'static MessageField

pub const TYPE_ARGUMENTS_FIELD: &'static MessageField

pub const ARGUMENTS_FIELD: &'static MessageField

impl MoveCall

pub fn path_builder() -> MoveCallFieldPathBuilder

impl MoveCall

pub const fn const_default() -> Self

pub fn package_opt_mut(&mut self) -> Option<&mut String>

If package is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn package_mut(&mut self) -> &mut String

Returns a mutable reference to package. If the field is unset, it is first initialized with the default value.

pub fn package_opt(&self) -> Option<&str>

If package is set, returns Some with the value; otherwise returns None.

pub fn set_package<T: Into<String>>(&mut self, field: T)

Sets package with the provided value.

pub fn with_package<T: Into<String>>(self, field: T) -> Self

Sets package with the provided value.

pub fn module_opt_mut(&mut self) -> Option<&mut String>

If module is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn module_mut(&mut self) -> &mut String

Returns a mutable reference to module. If the field is unset, it is first initialized with the default value.

pub fn module_opt(&self) -> Option<&str>

If module is set, returns Some with the value; otherwise returns None.

pub fn set_module<T: Into<String>>(&mut self, field: T)

Sets module with the provided value.

pub fn with_module<T: Into<String>>(self, field: T) -> Self

Sets module with the provided value.

pub fn function_opt_mut(&mut self) -> Option<&mut String>

If function is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn function_mut(&mut self) -> &mut String

Returns a mutable reference to function. If the field is unset, it is first initialized with the default value.

pub fn function_opt(&self) -> Option<&str>

If function is set, returns Some with the value; otherwise returns None.

pub fn set_function<T: Into<String>>(&mut self, field: T)

Sets function with the provided value.

pub fn with_function<T: Into<String>>(self, field: T) -> Self

Sets function with the provided value.

pub fn type_arguments(&self) -> &[String]

Returns the value of type_arguments, or the default value if type_arguments is unset.

pub fn type_arguments_mut(&mut self) -> &mut Vec<String>

Returns a mutable reference to type_arguments. If the field is unset, it is first initialized with the default value.

pub fn set_type_arguments(&mut self, field: Vec<String>)

Sets type_arguments with the provided value.

pub fn with_type_arguments(self, field: Vec<String>) -> Self

Sets type_arguments with the provided value.

pub fn arguments(&self) -> &[Argument]

Returns the value of arguments, or the default value if arguments is unset.

pub fn arguments_mut(&mut self) -> &mut Vec<Argument>

Returns a mutable reference to arguments. If the field is unset, it is first initialized with the default value.

pub fn set_arguments(&mut self, field: Vec<Argument>)

Sets arguments with the provided value.

pub fn with_arguments(self, field: Vec<Argument>) -> Self

Sets arguments with the provided value.

impl MoveCall

pub fn package(&self) -> &str

Returns the value of package, or the default value if package is unset.

pub fn module(&self) -> &str

Returns the value of module, or the default value if module is unset.

pub fn function(&self) -> &str

Returns the value of function, or the default value if function is unset.