sui_rpc::proto::sui::rpc::v2

Struct Object

#[non_exhaustive]
pub struct Object {

    pub bcs: Option<Bcs>,
    pub object_id: Option<String>,
    pub version: Option<u64>,
    pub digest: Option<String>,
    pub owner: Option<Owner>,
    pub object_type: Option<String>,
    pub has_public_transfer: Option<bool>,
    pub contents: Option<Bcs>,
    pub package: Option<Package>,
    pub previous_transaction: Option<String>,
    pub storage_rebate: Option<u64>,
    pub json: Option<Box<Value>>,
    pub balance: Option<u64>,
    pub display: Option<Box<Display>>,

}

An object on the Sui blockchain.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

bcs: Option<Bcs>

This Object serialized as BCS.

object_id: Option<String>

ObjectId for this object.

version: Option<u64>

Version of the object.

digest: Option<String>

The digest of this Object.

owner: Option<Owner>

Owner of the object.

object_type: Option<String>

The type of this object.

This will be ‘package’ for packages and a StructTag for move structs.

has_public_transfer: Option<bool>

DEPRECATED this field is no longer used to determine whether a tx can transfer this object. Instead, it is always calculated from the objects type when loaded in execution.

Only set for Move structs

contents: Option<Bcs>

BCS bytes of a Move struct value.

Only set for Move structs

package: Option<Package>

Package information for Move Packages

previous_transaction: Option<String>

The digest of the transaction that created or last mutated this object

storage_rebate: Option<u64>

The amount of SUI to rebate if this object gets deleted. This number is re-calculated each time the object is mutated based on the present storage gas price.

json: Option<Box<Value>>

JSON rendering of the object.

balance: Option<u64>

Current balance if this object is a 0x2::coin::Coin<T>

display: Option<Box<Display>>

JSON rendering of the object based on an on-chain template. This will not be set if the value’s type does not have an associated Display template.

Implementations

impl Object

pub const BCS_FIELD: &'static MessageField

pub const OBJECT_ID_FIELD: &'static MessageField

pub const VERSION_FIELD: &'static MessageField

pub const DIGEST_FIELD: &'static MessageField

pub const OWNER_FIELD: &'static MessageField

pub const OBJECT_TYPE_FIELD: &'static MessageField

pub const HAS_PUBLIC_TRANSFER_FIELD: &'static MessageField

pub const CONTENTS_FIELD: &'static MessageField

pub const PACKAGE_FIELD: &'static MessageField

pub const PREVIOUS_TRANSACTION_FIELD: &'static MessageField

pub const STORAGE_REBATE_FIELD: &'static MessageField

pub const JSON_FIELD: &'static MessageField

pub const BALANCE_FIELD: &'static MessageField

pub const DISPLAY_FIELD: &'static MessageField

impl Object

pub fn path_builder() -> ObjectFieldPathBuilder

impl Object

pub const fn const_default() -> Self

pub fn bcs(&self) -> &Bcs

Returns the value of bcs, or the default value if bcs is unset.

pub fn bcs_opt_mut(&mut self) -> Option<&mut Bcs>

If bcs is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn bcs_mut(&mut self) -> &mut Bcs

Returns a mutable reference to bcs. If the field is unset, it is first initialized with the default value.

pub fn bcs_opt(&self) -> Option<&Bcs>

If bcs is set, returns Some with the value; otherwise returns None.

pub fn set_bcs<T: Into<Bcs>>(&mut self, field: T)

Sets bcs with the provided value.

pub fn with_bcs<T: Into<Bcs>>(self, field: T) -> Self

Sets bcs with the provided value.

pub fn object_id_opt_mut(&mut self) -> Option<&mut String>

If object_id is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn object_id_mut(&mut self) -> &mut String

Returns a mutable reference to object_id. If the field is unset, it is first initialized with the default value.

pub fn object_id_opt(&self) -> Option<&str>

If object_id is set, returns Some with the value; otherwise returns None.

pub fn set_object_id<T: Into<String>>(&mut self, field: T)

Sets object_id with the provided value.

pub fn with_object_id<T: Into<String>>(self, field: T) -> Self

Sets object_id with the provided value.

pub fn version_opt_mut(&mut self) -> Option<&mut u64>

If version is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn version_mut(&mut self) -> &mut u64

Returns a mutable reference to version. If the field is unset, it is first initialized with the default value.

pub fn version_opt(&self) -> Option<u64>

If version is set, returns Some with the value; otherwise returns None.

pub fn set_version(&mut self, field: u64)

Sets version with the provided value.

pub fn with_version(self, field: u64) -> Self

Sets version with the provided value.

pub fn digest_opt_mut(&mut self) -> Option<&mut String>

If digest is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn digest_mut(&mut self) -> &mut String

Returns a mutable reference to digest. If the field is unset, it is first initialized with the default value.

pub fn digest_opt(&self) -> Option<&str>

If digest is set, returns Some with the value; otherwise returns None.

pub fn set_digest<T: Into<String>>(&mut self, field: T)

Sets digest with the provided value.

pub fn with_digest<T: Into<String>>(self, field: T) -> Self

Sets digest with the provided value.

pub fn owner(&self) -> &Owner

Returns the value of owner, or the default value if owner is unset.

pub fn owner_opt_mut(&mut self) -> Option<&mut Owner>

If owner is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn owner_mut(&mut self) -> &mut Owner

Returns a mutable reference to owner. If the field is unset, it is first initialized with the default value.

pub fn owner_opt(&self) -> Option<&Owner>

If owner is set, returns Some with the value; otherwise returns None.

pub fn set_owner<T: Into<Owner>>(&mut self, field: T)

Sets owner with the provided value.

pub fn with_owner<T: Into<Owner>>(self, field: T) -> Self

Sets owner with the provided value.

pub fn object_type_opt_mut(&mut self) -> Option<&mut String>

If object_type is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn object_type_mut(&mut self) -> &mut String

Returns a mutable reference to object_type. If the field is unset, it is first initialized with the default value.

pub fn object_type_opt(&self) -> Option<&str>

If object_type is set, returns Some with the value; otherwise returns None.

pub fn set_object_type<T: Into<String>>(&mut self, field: T)

Sets object_type with the provided value.

pub fn with_object_type<T: Into<String>>(self, field: T) -> Self

Sets object_type with the provided value.

pub fn has_public_transfer_opt_mut(&mut self) -> Option<&mut bool>

If has_public_transfer is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn has_public_transfer_mut(&mut self) -> &mut bool

Returns a mutable reference to has_public_transfer. If the field is unset, it is first initialized with the default value.

pub fn has_public_transfer_opt(&self) -> Option<bool>

If has_public_transfer is set, returns Some with the value; otherwise returns None.

pub fn set_has_public_transfer(&mut self, field: bool)

Sets has_public_transfer with the provided value.

pub fn with_has_public_transfer(self, field: bool) -> Self

Sets has_public_transfer with the provided value.

pub fn contents(&self) -> &Bcs

Returns the value of contents, or the default value if contents is unset.

pub fn contents_opt_mut(&mut self) -> Option<&mut Bcs>

If contents is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn contents_mut(&mut self) -> &mut Bcs

Returns a mutable reference to contents. If the field is unset, it is first initialized with the default value.

pub fn contents_opt(&self) -> Option<&Bcs>

If contents is set, returns Some with the value; otherwise returns None.

pub fn set_contents<T: Into<Bcs>>(&mut self, field: T)

Sets contents with the provided value.

pub fn with_contents<T: Into<Bcs>>(self, field: T) -> Self

Sets contents with the provided value.

pub fn package(&self) -> &Package

Returns the value of package, or the default value if package is unset.

pub fn package_opt_mut(&mut self) -> Option<&mut Package>

If package is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn package_mut(&mut self) -> &mut Package

Returns a mutable reference to package. If the field is unset, it is first initialized with the default value.

pub fn package_opt(&self) -> Option<&Package>

If package is set, returns Some with the value; otherwise returns None.

pub fn set_package<T: Into<Package>>(&mut self, field: T)

Sets package with the provided value.

pub fn with_package<T: Into<Package>>(self, field: T) -> Self

Sets package with the provided value.

pub fn previous_transaction_opt_mut(&mut self) -> Option<&mut String>

If previous_transaction is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn previous_transaction_mut(&mut self) -> &mut String

Returns a mutable reference to previous_transaction. If the field is unset, it is first initialized with the default value.

pub fn previous_transaction_opt(&self) -> Option<&str>

If previous_transaction is set, returns Some with the value; otherwise returns None.

pub fn set_previous_transaction<T: Into<String>>(&mut self, field: T)

Sets previous_transaction with the provided value.

pub fn with_previous_transaction<T: Into<String>>(self, field: T) -> Self

Sets previous_transaction with the provided value.

pub fn storage_rebate_opt_mut(&mut self) -> Option<&mut u64>

If storage_rebate is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn storage_rebate_mut(&mut self) -> &mut u64

Returns a mutable reference to storage_rebate. If the field is unset, it is first initialized with the default value.

pub fn storage_rebate_opt(&self) -> Option<u64>

If storage_rebate is set, returns Some with the value; otherwise returns None.

pub fn set_storage_rebate(&mut self, field: u64)

Sets storage_rebate with the provided value.

pub fn with_storage_rebate(self, field: u64) -> Self

Sets storage_rebate with the provided value.

pub fn json_opt_mut(&mut self) -> Option<&mut Value>

If json is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn json_mut(&mut self) -> &mut Value

Returns a mutable reference to json. If the field is unset, it is first initialized with the default value.

pub fn json_opt(&self) -> Option<&Value>

If json is set, returns Some with the value; otherwise returns None.

pub fn set_json<T: Into<Value>>(&mut self, field: T)

Sets json with the provided value.

pub fn with_json<T: Into<Value>>(self, field: T) -> Self

Sets json with the provided value.

pub fn balance_opt_mut(&mut self) -> Option<&mut u64>

If balance is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn balance_mut(&mut self) -> &mut u64

Returns a mutable reference to balance. If the field is unset, it is first initialized with the default value.

pub fn balance_opt(&self) -> Option<u64>

If balance is set, returns Some with the value; otherwise returns None.

pub fn set_balance(&mut self, field: u64)

Sets balance with the provided value.

pub fn with_balance(self, field: u64) -> Self

Sets balance with the provided value.

pub fn display(&self) -> &Display

Returns the value of display, or the default value if display is unset.

pub fn display_opt_mut(&mut self) -> Option<&mut Display>

If display is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn display_mut(&mut self) -> &mut Display

Returns a mutable reference to display. If the field is unset, it is first initialized with the default value.

pub fn display_opt(&self) -> Option<&Display>

If display is set, returns Some with the value; otherwise returns None.

pub fn set_display<T: Into<Display>>(&mut self, field: T)

Sets display with the provided value.

pub fn with_display<T: Into<Display>>(self, field: T) -> Self

Sets display with the provided value.

impl Object

pub fn object_reference(&self) -> ObjectReference

impl Object

pub fn object_id(&self) -> &str

Returns the value of object_id, or the default value if object_id is unset.

pub fn version(&self) -> u64

Returns the value of version, or the default value if version is unset.

pub fn digest(&self) -> &str

Returns the value of digest, or the default value if digest is unset.

pub fn object_type(&self) -> &str

Returns the value of object_type, or the default value if object_type is unset.

pub fn has_public_transfer(&self) -> bool

Returns the value of has_public_transfer, or the default value if has_public_transfer is unset.

pub fn previous_transaction(&self) -> &str

Returns the value of previous_transaction, or the default value if previous_transaction is unset.

pub fn storage_rebate(&self) -> u64

Returns the value of storage_rebate, or the default value if storage_rebate is unset.

pub fn balance(&self) -> u64

Returns the value of balance, or the default value if balance is unset.