sui_rpc::proto::sui::rpc::v2

Struct SystemParameters

#[non_exhaustive]
pub struct SystemParameters {
    pub epoch_duration_ms: Option<u64>,
    pub stake_subsidy_start_epoch: Option<u64>,
    pub min_validator_count: Option<u64>,
    pub max_validator_count: Option<u64>,
    pub min_validator_joining_stake: Option<u64>,
    pub validator_low_stake_threshold: Option<u64>,
    pub validator_very_low_stake_threshold: Option<u64>,
    pub validator_low_stake_grace_period: Option<u64>,
    pub extra_fields: Option<MoveTable>,
}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

epoch_duration_ms: Option<u64>

The duration of an epoch, in milliseconds.

stake_subsidy_start_epoch: Option<u64>

The starting epoch in which stake subsidies start being paid out

min_validator_count: Option<u64>

Minimum number of active validators at any moment.

max_validator_count: Option<u64>

Maximum number of active validators at any moment. We do not allow the number of validators in any epoch to go above this.

min_validator_joining_stake: Option<u64>

Deprecated. Lower-bound on the amount of stake required to become a validator.

validator_low_stake_threshold: Option<u64>

Deprecated. Validators with stake amount below validator_low_stake_threshold are considered to have low stake and will be escorted out of the validator set after being below this threshold for more than validator_low_stake_grace_period number of epochs.

validator_very_low_stake_threshold: Option<u64>

Deprecated. Validators with stake below validator_very_low_stake_threshold will be removed immediately at epoch change, no grace period.

validator_low_stake_grace_period: Option<u64>

A validator can have stake below validator_low_stake_threshold for this many epochs before being kicked out.

extra_fields: Option<MoveTable>

Any extra fields that are not defined statically.

Implementations

impl SystemParameters

pub const EPOCH_DURATION_MS_FIELD: &'static MessageField

pub const STAKE_SUBSIDY_START_EPOCH_FIELD: &'static MessageField

pub const MIN_VALIDATOR_COUNT_FIELD: &'static MessageField

pub const MAX_VALIDATOR_COUNT_FIELD: &'static MessageField

pub const MIN_VALIDATOR_JOINING_STAKE_FIELD: &'static MessageField

pub const VALIDATOR_LOW_STAKE_THRESHOLD_FIELD: &'static MessageField

pub const VALIDATOR_VERY_LOW_STAKE_THRESHOLD_FIELD: &'static MessageField

pub const VALIDATOR_LOW_STAKE_GRACE_PERIOD_FIELD: &'static MessageField

pub const EXTRA_FIELDS_FIELD: &'static MessageField

impl SystemParameters

pub fn path_builder() -> SystemParametersFieldPathBuilder

impl SystemParameters

pub const fn const_default() -> Self

pub fn epoch_duration_ms_opt_mut(&mut self) -> Option<&mut u64>

If epoch_duration_ms is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn epoch_duration_ms_mut(&mut self) -> &mut u64

Returns a mutable reference to epoch_duration_ms. If the field is unset, it is first initialized with the default value.

pub fn epoch_duration_ms_opt(&self) -> Option<u64>

If epoch_duration_ms is set, returns Some with the value; otherwise returns None.

pub fn set_epoch_duration_ms(&mut self, field: u64)

Sets epoch_duration_ms with the provided value.

pub fn with_epoch_duration_ms(self, field: u64) -> Self

Sets epoch_duration_ms with the provided value.

pub fn stake_subsidy_start_epoch_opt_mut(&mut self) -> Option<&mut u64>

If stake_subsidy_start_epoch is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn stake_subsidy_start_epoch_mut(&mut self) -> &mut u64

Returns a mutable reference to stake_subsidy_start_epoch. If the field is unset, it is first initialized with the default value.

pub fn stake_subsidy_start_epoch_opt(&self) -> Option<u64>

If stake_subsidy_start_epoch is set, returns Some with the value; otherwise returns None.

pub fn set_stake_subsidy_start_epoch(&mut self, field: u64)

Sets stake_subsidy_start_epoch with the provided value.

pub fn with_stake_subsidy_start_epoch(self, field: u64) -> Self

Sets stake_subsidy_start_epoch with the provided value.

pub fn min_validator_count_opt_mut(&mut self) -> Option<&mut u64>

If min_validator_count is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn min_validator_count_mut(&mut self) -> &mut u64

Returns a mutable reference to min_validator_count. If the field is unset, it is first initialized with the default value.

pub fn min_validator_count_opt(&self) -> Option<u64>

If min_validator_count is set, returns Some with the value; otherwise returns None.

pub fn set_min_validator_count(&mut self, field: u64)

Sets min_validator_count with the provided value.

pub fn with_min_validator_count(self, field: u64) -> Self

Sets min_validator_count with the provided value.

pub fn max_validator_count_opt_mut(&mut self) -> Option<&mut u64>

If max_validator_count is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn max_validator_count_mut(&mut self) -> &mut u64

Returns a mutable reference to max_validator_count. If the field is unset, it is first initialized with the default value.

pub fn max_validator_count_opt(&self) -> Option<u64>

If max_validator_count is set, returns Some with the value; otherwise returns None.

pub fn set_max_validator_count(&mut self, field: u64)

Sets max_validator_count with the provided value.

pub fn with_max_validator_count(self, field: u64) -> Self

Sets max_validator_count with the provided value.

pub fn min_validator_joining_stake_opt_mut(&mut self) -> Option<&mut u64>

If min_validator_joining_stake is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn min_validator_joining_stake_mut(&mut self) -> &mut u64

Returns a mutable reference to min_validator_joining_stake. If the field is unset, it is first initialized with the default value.

pub fn min_validator_joining_stake_opt(&self) -> Option<u64>

If min_validator_joining_stake is set, returns Some with the value; otherwise returns None.

pub fn set_min_validator_joining_stake(&mut self, field: u64)

Sets min_validator_joining_stake with the provided value.

pub fn with_min_validator_joining_stake(self, field: u64) -> Self

Sets min_validator_joining_stake with the provided value.

pub fn validator_low_stake_threshold_opt_mut(&mut self) -> Option<&mut u64>

If validator_low_stake_threshold is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn validator_low_stake_threshold_mut(&mut self) -> &mut u64

Returns a mutable reference to validator_low_stake_threshold. If the field is unset, it is first initialized with the default value.

pub fn validator_low_stake_threshold_opt(&self) -> Option<u64>

If validator_low_stake_threshold is set, returns Some with the value; otherwise returns None.

pub fn set_validator_low_stake_threshold(&mut self, field: u64)

Sets validator_low_stake_threshold with the provided value.

pub fn with_validator_low_stake_threshold(self, field: u64) -> Self

Sets validator_low_stake_threshold with the provided value.

pub fn validator_very_low_stake_threshold_opt_mut(&mut self) -> Option<&mut u64>

If validator_very_low_stake_threshold is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn validator_very_low_stake_threshold_mut(&mut self) -> &mut u64

Returns a mutable reference to validator_very_low_stake_threshold. If the field is unset, it is first initialized with the default value.

pub fn validator_very_low_stake_threshold_opt(&self) -> Option<u64>

If validator_very_low_stake_threshold is set, returns Some with the value; otherwise returns None.

pub fn set_validator_very_low_stake_threshold(&mut self, field: u64)

Sets validator_very_low_stake_threshold with the provided value.

pub fn with_validator_very_low_stake_threshold(self, field: u64) -> Self

Sets validator_very_low_stake_threshold with the provided value.

pub fn validator_low_stake_grace_period_opt_mut(&mut self) -> Option<&mut u64>

If validator_low_stake_grace_period is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn validator_low_stake_grace_period_mut(&mut self) -> &mut u64

Returns a mutable reference to validator_low_stake_grace_period. If the field is unset, it is first initialized with the default value.

pub fn validator_low_stake_grace_period_opt(&self) -> Option<u64>

If validator_low_stake_grace_period is set, returns Some with the value; otherwise returns None.

pub fn set_validator_low_stake_grace_period(&mut self, field: u64)

Sets validator_low_stake_grace_period with the provided value.

pub fn with_validator_low_stake_grace_period(self, field: u64) -> Self

Sets validator_low_stake_grace_period with the provided value.

pub fn extra_fields(&self) -> &MoveTable

Returns the value of extra_fields, or the default value if extra_fields is unset.

pub fn extra_fields_opt_mut(&mut self) -> Option<&mut MoveTable>

If extra_fields is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn extra_fields_mut(&mut self) -> &mut MoveTable

Returns a mutable reference to extra_fields. If the field is unset, it is first initialized with the default value.

pub fn extra_fields_opt(&self) -> Option<&MoveTable>

If extra_fields is set, returns Some with the value; otherwise returns None.

pub fn set_extra_fields<T: Into<MoveTable>>(&mut self, field: T)

Sets extra_fields with the provided value.

pub fn with_extra_fields<T: Into<MoveTable>>(self, field: T) -> Self

Sets extra_fields with the provided value.

impl SystemParameters

pub fn epoch_duration_ms(&self) -> u64

Returns the value of epoch_duration_ms, or the default value if epoch_duration_ms is unset.

pub fn stake_subsidy_start_epoch(&self) -> u64

Returns the value of stake_subsidy_start_epoch, or the default value if stake_subsidy_start_epoch is unset.

pub fn min_validator_count(&self) -> u64

Returns the value of min_validator_count, or the default value if min_validator_count is unset.

pub fn max_validator_count(&self) -> u64

Returns the value of max_validator_count, or the default value if max_validator_count is unset.

pub fn min_validator_joining_stake(&self) -> u64

Returns the value of min_validator_joining_stake, or the default value if min_validator_joining_stake is unset.

pub fn validator_low_stake_threshold(&self) -> u64

Returns the value of validator_low_stake_threshold, or the default value if validator_low_stake_threshold is unset.

pub fn validator_very_low_stake_threshold(&self) -> u64

Returns the value of validator_very_low_stake_threshold, or the default value if validator_very_low_stake_threshold is unset.

pub fn validator_low_stake_grace_period(&self) -> u64

Returns the value of validator_low_stake_grace_period, or the default value if validator_low_stake_grace_period is unset.