sui_rpc::proto::sui::rpc::v2

Struct Validator

#[non_exhaustive]
pub struct Validator {

    pub name: Option<String>,
    pub address: Option<String>,
    pub description: Option<String>,
    pub image_url: Option<String>,
    pub project_url: Option<String>,
    pub protocol_public_key: Option<Bytes>,
    pub proof_of_possession: Option<Bytes>,
    pub network_public_key: Option<Bytes>,
    pub worker_public_key: Option<Bytes>,
    pub network_address: Option<String>,
    pub p2p_address: Option<String>,
    pub primary_address: Option<String>,
    pub worker_address: Option<String>,
    pub next_epoch_protocol_public_key: Option<Bytes>,
    pub next_epoch_proof_of_possession: Option<Bytes>,
    pub next_epoch_network_public_key: Option<Bytes>,
    pub next_epoch_worker_public_key: Option<Bytes>,
    pub next_epoch_network_address: Option<String>,
    pub next_epoch_p2p_address: Option<String>,
    pub next_epoch_primary_address: Option<String>,
    pub next_epoch_worker_address: Option<String>,
    pub metadata_extra_fields: Option<MoveTable>,
    pub voting_power: Option<u64>,
    pub operation_cap_id: Option<String>,
    pub gas_price: Option<u64>,
    pub staking_pool: Option<StakingPool>,
    pub commission_rate: Option<u64>,
    pub next_epoch_stake: Option<u64>,
    pub next_epoch_gas_price: Option<u64>,
    pub next_epoch_commission_rate: Option<u64>,
    pub extra_fields: Option<MoveTable>,

}

Definition of a Validator in the system contracts

Note: fields of ValidatorMetadata are flattened into this type

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

name: Option<String>

A unique human-readable name of this validator.

address: Option<String>

The Sui Address of the validator. This is the sender that created the Validator object, and also the address to send validator/coins to during withdraws.

description: Option<String>image_url: Option<String>project_url: Option<String>protocol_public_key: Option<Bytes>

The public key bytes corresponding to the private key that the validator holds to sign transactions. For now, this is the same as AuthorityName.

proof_of_possession: Option<Bytes>

This is a proof that the validator has ownership of the protocol private key

network_public_key: Option<Bytes>

The public key bytes corresponding to the private key that the validator uses to establish TLS connections

worker_public_key: Option<Bytes>

The public key bytes corresponding to the Narwhal Worker

network_address: Option<String>

The network address of the validator (could also contain extra info such as port, DNS and etc.).

p2p_address: Option<String>

The address of the validator used for p2p activities such as state sync (could also contain extra info such as port, DNS and etc.).

primary_address: Option<String>

The address of the narwhal primary

worker_address: Option<String>

The address of the narwhal worker

next_epoch_protocol_public_key: Option<Bytes>next_epoch_proof_of_possession: Option<Bytes>next_epoch_network_public_key: Option<Bytes>next_epoch_worker_public_key: Option<Bytes>next_epoch_network_address: Option<String>next_epoch_p2p_address: Option<String>next_epoch_primary_address: Option<String>next_epoch_worker_address: Option<String>metadata_extra_fields: Option<MoveTable>

Any extra fields that’s not defined statically in the ValidatorMetadata struct

voting_power: Option<u64>

The voting power of this validator, which might be different from its stake amount.

operation_cap_id: Option<String>

The ID of this validator’s current valid UnverifiedValidatorOperationCap

gas_price: Option<u64>

Gas price quote, updated only at end of epoch.

staking_pool: Option<StakingPool>

Staking pool for this validator.

commission_rate: Option<u64>

Commission rate of the validator, in basis point.

next_epoch_stake: Option<u64>

Total amount of stake that would be active in the next epoch.

next_epoch_gas_price: Option<u64>

This validator’s gas price quote for the next epoch.

next_epoch_commission_rate: Option<u64>

The commission rate of the validator starting the next epoch, in basis point.

extra_fields: Option<MoveTable>

Any extra fields that’s not defined statically.

Implementations

impl Validator

pub const NAME_FIELD: &'static MessageField

pub const ADDRESS_FIELD: &'static MessageField

pub const DESCRIPTION_FIELD: &'static MessageField

pub const IMAGE_URL_FIELD: &'static MessageField

pub const PROJECT_URL_FIELD: &'static MessageField

pub const PROTOCOL_PUBLIC_KEY_FIELD: &'static MessageField

pub const PROOF_OF_POSSESSION_FIELD: &'static MessageField

pub const NETWORK_PUBLIC_KEY_FIELD: &'static MessageField

pub const WORKER_PUBLIC_KEY_FIELD: &'static MessageField

pub const NETWORK_ADDRESS_FIELD: &'static MessageField

pub const P2P_ADDRESS_FIELD: &'static MessageField

pub const PRIMARY_ADDRESS_FIELD: &'static MessageField

pub const WORKER_ADDRESS_FIELD: &'static MessageField

pub const NEXT_EPOCH_PROTOCOL_PUBLIC_KEY_FIELD: &'static MessageField

pub const NEXT_EPOCH_PROOF_OF_POSSESSION_FIELD: &'static MessageField

pub const NEXT_EPOCH_NETWORK_PUBLIC_KEY_FIELD: &'static MessageField

pub const NEXT_EPOCH_WORKER_PUBLIC_KEY_FIELD: &'static MessageField

pub const NEXT_EPOCH_NETWORK_ADDRESS_FIELD: &'static MessageField

pub const NEXT_EPOCH_P2P_ADDRESS_FIELD: &'static MessageField

pub const NEXT_EPOCH_PRIMARY_ADDRESS_FIELD: &'static MessageField

pub const NEXT_EPOCH_WORKER_ADDRESS_FIELD: &'static MessageField

pub const METADATA_EXTRA_FIELDS_FIELD: &'static MessageField

pub const VOTING_POWER_FIELD: &'static MessageField

pub const OPERATION_CAP_ID_FIELD: &'static MessageField

pub const GAS_PRICE_FIELD: &'static MessageField

pub const STAKING_POOL_FIELD: &'static MessageField

pub const COMMISSION_RATE_FIELD: &'static MessageField

pub const NEXT_EPOCH_STAKE_FIELD: &'static MessageField

pub const NEXT_EPOCH_GAS_PRICE_FIELD: &'static MessageField

pub const NEXT_EPOCH_COMMISSION_RATE_FIELD: &'static MessageField

pub const EXTRA_FIELDS_FIELD: &'static MessageField

impl Validator

pub fn path_builder() -> ValidatorFieldPathBuilder

impl Validator

pub const fn const_default() -> Self

pub fn name_opt_mut(&mut self) -> Option<&mut String>

If name is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn name_mut(&mut self) -> &mut String

Returns a mutable reference to name. If the field is unset, it is first initialized with the default value.

pub fn name_opt(&self) -> Option<&str>

If name is set, returns Some with the value; otherwise returns None.

pub fn set_name<T: Into<String>>(&mut self, field: T)

Sets name with the provided value.

pub fn with_name<T: Into<String>>(self, field: T) -> Self

Sets name with the provided value.

pub fn address_opt_mut(&mut self) -> Option<&mut String>

If address is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn address_mut(&mut self) -> &mut String

Returns a mutable reference to address. If the field is unset, it is first initialized with the default value.

pub fn address_opt(&self) -> Option<&str>

If address is set, returns Some with the value; otherwise returns None.

pub fn set_address<T: Into<String>>(&mut self, field: T)

Sets address with the provided value.

pub fn with_address<T: Into<String>>(self, field: T) -> Self

Sets address with the provided value.

pub fn description_opt_mut(&mut self) -> Option<&mut String>

If description is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn description_mut(&mut self) -> &mut String

Returns a mutable reference to description. If the field is unset, it is first initialized with the default value.

pub fn description_opt(&self) -> Option<&str>

If description is set, returns Some with the value; otherwise returns None.

pub fn set_description<T: Into<String>>(&mut self, field: T)

Sets description with the provided value.

pub fn with_description<T: Into<String>>(self, field: T) -> Self

Sets description with the provided value.

pub fn image_url_opt_mut(&mut self) -> Option<&mut String>

If image_url is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn image_url_mut(&mut self) -> &mut String

Returns a mutable reference to image_url. If the field is unset, it is first initialized with the default value.

pub fn image_url_opt(&self) -> Option<&str>

If image_url is set, returns Some with the value; otherwise returns None.

pub fn set_image_url<T: Into<String>>(&mut self, field: T)

Sets image_url with the provided value.

pub fn with_image_url<T: Into<String>>(self, field: T) -> Self

Sets image_url with the provided value.

pub fn project_url_opt_mut(&mut self) -> Option<&mut String>

If project_url is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn project_url_mut(&mut self) -> &mut String

Returns a mutable reference to project_url. If the field is unset, it is first initialized with the default value.

pub fn project_url_opt(&self) -> Option<&str>

If project_url is set, returns Some with the value; otherwise returns None.

pub fn set_project_url<T: Into<String>>(&mut self, field: T)

Sets project_url with the provided value.

pub fn with_project_url<T: Into<String>>(self, field: T) -> Self

Sets project_url with the provided value.

pub fn protocol_public_key_opt(&self) -> Option<&[u8]>

If protocol_public_key is set, returns Some with the value; otherwise returns None.

pub fn set_protocol_public_key<T: Into<Bytes>>(&mut self, field: T)

Sets protocol_public_key with the provided value.

pub fn with_protocol_public_key<T: Into<Bytes>>(self, field: T) -> Self

Sets protocol_public_key with the provided value.

pub fn proof_of_possession_opt(&self) -> Option<&[u8]>

If proof_of_possession is set, returns Some with the value; otherwise returns None.

pub fn set_proof_of_possession<T: Into<Bytes>>(&mut self, field: T)

Sets proof_of_possession with the provided value.

pub fn with_proof_of_possession<T: Into<Bytes>>(self, field: T) -> Self

Sets proof_of_possession with the provided value.

pub fn network_public_key_opt(&self) -> Option<&[u8]>

If network_public_key is set, returns Some with the value; otherwise returns None.

pub fn set_network_public_key<T: Into<Bytes>>(&mut self, field: T)

Sets network_public_key with the provided value.

pub fn with_network_public_key<T: Into<Bytes>>(self, field: T) -> Self

Sets network_public_key with the provided value.

pub fn worker_public_key_opt(&self) -> Option<&[u8]>

If worker_public_key is set, returns Some with the value; otherwise returns None.

pub fn set_worker_public_key<T: Into<Bytes>>(&mut self, field: T)

Sets worker_public_key with the provided value.

pub fn with_worker_public_key<T: Into<Bytes>>(self, field: T) -> Self

Sets worker_public_key with the provided value.

pub fn network_address_opt_mut(&mut self) -> Option<&mut String>

If network_address is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn network_address_mut(&mut self) -> &mut String

Returns a mutable reference to network_address. If the field is unset, it is first initialized with the default value.

pub fn network_address_opt(&self) -> Option<&str>

If network_address is set, returns Some with the value; otherwise returns None.

pub fn set_network_address<T: Into<String>>(&mut self, field: T)

Sets network_address with the provided value.

pub fn with_network_address<T: Into<String>>(self, field: T) -> Self

Sets network_address with the provided value.

pub fn p2p_address_opt_mut(&mut self) -> Option<&mut String>

If p2p_address is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn p2p_address_mut(&mut self) -> &mut String

Returns a mutable reference to p2p_address. If the field is unset, it is first initialized with the default value.

pub fn p2p_address_opt(&self) -> Option<&str>

If p2p_address is set, returns Some with the value; otherwise returns None.

pub fn set_p2p_address<T: Into<String>>(&mut self, field: T)

Sets p2p_address with the provided value.

pub fn with_p2p_address<T: Into<String>>(self, field: T) -> Self

Sets p2p_address with the provided value.

pub fn primary_address_opt_mut(&mut self) -> Option<&mut String>

If primary_address is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn primary_address_mut(&mut self) -> &mut String

Returns a mutable reference to primary_address. If the field is unset, it is first initialized with the default value.

pub fn primary_address_opt(&self) -> Option<&str>

If primary_address is set, returns Some with the value; otherwise returns None.

pub fn set_primary_address<T: Into<String>>(&mut self, field: T)

Sets primary_address with the provided value.

pub fn with_primary_address<T: Into<String>>(self, field: T) -> Self

Sets primary_address with the provided value.

pub fn worker_address_opt_mut(&mut self) -> Option<&mut String>

If worker_address is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn worker_address_mut(&mut self) -> &mut String

Returns a mutable reference to worker_address. If the field is unset, it is first initialized with the default value.

pub fn worker_address_opt(&self) -> Option<&str>

If worker_address is set, returns Some with the value; otherwise returns None.

pub fn set_worker_address<T: Into<String>>(&mut self, field: T)

Sets worker_address with the provided value.

pub fn with_worker_address<T: Into<String>>(self, field: T) -> Self

Sets worker_address with the provided value.

pub fn next_epoch_protocol_public_key_opt(&self) -> Option<&[u8]>

If next_epoch_protocol_public_key is set, returns Some with the value; otherwise returns None.

pub fn set_next_epoch_protocol_public_key<T: Into<Bytes>>(&mut self, field: T)

Sets next_epoch_protocol_public_key with the provided value.

pub fn with_next_epoch_protocol_public_key<T: Into<Bytes>>( self, field: T, ) -> Self

Sets next_epoch_protocol_public_key with the provided value.

pub fn next_epoch_proof_of_possession_opt(&self) -> Option<&[u8]>

If next_epoch_proof_of_possession is set, returns Some with the value; otherwise returns None.

pub fn set_next_epoch_proof_of_possession<T: Into<Bytes>>(&mut self, field: T)

Sets next_epoch_proof_of_possession with the provided value.

pub fn with_next_epoch_proof_of_possession<T: Into<Bytes>>( self, field: T, ) -> Self

Sets next_epoch_proof_of_possession with the provided value.

pub fn next_epoch_network_public_key_opt(&self) -> Option<&[u8]>

If next_epoch_network_public_key is set, returns Some with the value; otherwise returns None.

pub fn set_next_epoch_network_public_key<T: Into<Bytes>>(&mut self, field: T)

Sets next_epoch_network_public_key with the provided value.

pub fn with_next_epoch_network_public_key<T: Into<Bytes>>( self, field: T, ) -> Self

Sets next_epoch_network_public_key with the provided value.

pub fn next_epoch_worker_public_key_opt(&self) -> Option<&[u8]>

If next_epoch_worker_public_key is set, returns Some with the value; otherwise returns None.

pub fn set_next_epoch_worker_public_key<T: Into<Bytes>>(&mut self, field: T)

Sets next_epoch_worker_public_key with the provided value.

pub fn with_next_epoch_worker_public_key<T: Into<Bytes>>(self, field: T) -> Self

Sets next_epoch_worker_public_key with the provided value.

pub fn next_epoch_network_address_opt_mut(&mut self) -> Option<&mut String>

If next_epoch_network_address is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn next_epoch_network_address_mut(&mut self) -> &mut String

Returns a mutable reference to next_epoch_network_address. If the field is unset, it is first initialized with the default value.

pub fn next_epoch_network_address_opt(&self) -> Option<&str>

If next_epoch_network_address is set, returns Some with the value; otherwise returns None.

pub fn set_next_epoch_network_address<T: Into<String>>(&mut self, field: T)

Sets next_epoch_network_address with the provided value.

pub fn with_next_epoch_network_address<T: Into<String>>(self, field: T) -> Self

Sets next_epoch_network_address with the provided value.

pub fn next_epoch_p2p_address_opt_mut(&mut self) -> Option<&mut String>

If next_epoch_p2p_address is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn next_epoch_p2p_address_mut(&mut self) -> &mut String

Returns a mutable reference to next_epoch_p2p_address. If the field is unset, it is first initialized with the default value.

pub fn next_epoch_p2p_address_opt(&self) -> Option<&str>

If next_epoch_p2p_address is set, returns Some with the value; otherwise returns None.

pub fn set_next_epoch_p2p_address<T: Into<String>>(&mut self, field: T)

Sets next_epoch_p2p_address with the provided value.

pub fn with_next_epoch_p2p_address<T: Into<String>>(self, field: T) -> Self

Sets next_epoch_p2p_address with the provided value.

pub fn next_epoch_primary_address_opt_mut(&mut self) -> Option<&mut String>

If next_epoch_primary_address is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn next_epoch_primary_address_mut(&mut self) -> &mut String

Returns a mutable reference to next_epoch_primary_address. If the field is unset, it is first initialized with the default value.

pub fn next_epoch_primary_address_opt(&self) -> Option<&str>

If next_epoch_primary_address is set, returns Some with the value; otherwise returns None.

pub fn set_next_epoch_primary_address<T: Into<String>>(&mut self, field: T)

Sets next_epoch_primary_address with the provided value.

pub fn with_next_epoch_primary_address<T: Into<String>>(self, field: T) -> Self

Sets next_epoch_primary_address with the provided value.

pub fn next_epoch_worker_address_opt_mut(&mut self) -> Option<&mut String>

If next_epoch_worker_address is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn next_epoch_worker_address_mut(&mut self) -> &mut String

Returns a mutable reference to next_epoch_worker_address. If the field is unset, it is first initialized with the default value.

pub fn next_epoch_worker_address_opt(&self) -> Option<&str>

If next_epoch_worker_address is set, returns Some with the value; otherwise returns None.

pub fn set_next_epoch_worker_address<T: Into<String>>(&mut self, field: T)

Sets next_epoch_worker_address with the provided value.

pub fn with_next_epoch_worker_address<T: Into<String>>(self, field: T) -> Self

Sets next_epoch_worker_address with the provided value.

pub fn metadata_extra_fields(&self) -> &MoveTable

Returns the value of metadata_extra_fields, or the default value if metadata_extra_fields is unset.

pub fn metadata_extra_fields_opt_mut(&mut self) -> Option<&mut MoveTable>

If metadata_extra_fields is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn metadata_extra_fields_mut(&mut self) -> &mut MoveTable

Returns a mutable reference to metadata_extra_fields. If the field is unset, it is first initialized with the default value.

pub fn metadata_extra_fields_opt(&self) -> Option<&MoveTable>

If metadata_extra_fields is set, returns Some with the value; otherwise returns None.

pub fn set_metadata_extra_fields<T: Into<MoveTable>>(&mut self, field: T)

Sets metadata_extra_fields with the provided value.

pub fn with_metadata_extra_fields<T: Into<MoveTable>>(self, field: T) -> Self

Sets metadata_extra_fields with the provided value.

pub fn voting_power_opt_mut(&mut self) -> Option<&mut u64>

If voting_power is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn voting_power_mut(&mut self) -> &mut u64

Returns a mutable reference to voting_power. If the field is unset, it is first initialized with the default value.

pub fn voting_power_opt(&self) -> Option<u64>

If voting_power is set, returns Some with the value; otherwise returns None.

pub fn set_voting_power<T: Into<u64>>(&mut self, field: T)

Sets voting_power with the provided value.

pub fn with_voting_power<T: Into<u64>>(self, field: T) -> Self

Sets voting_power with the provided value.

pub fn operation_cap_id_opt_mut(&mut self) -> Option<&mut String>

If operation_cap_id is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn operation_cap_id_mut(&mut self) -> &mut String

Returns a mutable reference to operation_cap_id. If the field is unset, it is first initialized with the default value.

pub fn operation_cap_id_opt(&self) -> Option<&str>

If operation_cap_id is set, returns Some with the value; otherwise returns None.

pub fn set_operation_cap_id<T: Into<String>>(&mut self, field: T)

Sets operation_cap_id with the provided value.

pub fn with_operation_cap_id<T: Into<String>>(self, field: T) -> Self

Sets operation_cap_id with the provided value.

pub fn gas_price_opt_mut(&mut self) -> Option<&mut u64>

If gas_price is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn gas_price_mut(&mut self) -> &mut u64

Returns a mutable reference to gas_price. If the field is unset, it is first initialized with the default value.

pub fn gas_price_opt(&self) -> Option<u64>

If gas_price is set, returns Some with the value; otherwise returns None.

pub fn set_gas_price<T: Into<u64>>(&mut self, field: T)

Sets gas_price with the provided value.

pub fn with_gas_price<T: Into<u64>>(self, field: T) -> Self

Sets gas_price with the provided value.

pub fn staking_pool(&self) -> &StakingPool

Returns the value of staking_pool, or the default value if staking_pool is unset.

pub fn staking_pool_opt_mut(&mut self) -> Option<&mut StakingPool>

If staking_pool is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn staking_pool_mut(&mut self) -> &mut StakingPool

Returns a mutable reference to staking_pool. If the field is unset, it is first initialized with the default value.

pub fn staking_pool_opt(&self) -> Option<&StakingPool>

If staking_pool is set, returns Some with the value; otherwise returns None.

pub fn set_staking_pool<T: Into<StakingPool>>(&mut self, field: T)

Sets staking_pool with the provided value.

pub fn with_staking_pool<T: Into<StakingPool>>(self, field: T) -> Self

Sets staking_pool with the provided value.

pub fn commission_rate_opt_mut(&mut self) -> Option<&mut u64>

If commission_rate is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn commission_rate_mut(&mut self) -> &mut u64

Returns a mutable reference to commission_rate. If the field is unset, it is first initialized with the default value.

pub fn commission_rate_opt(&self) -> Option<u64>

If commission_rate is set, returns Some with the value; otherwise returns None.

pub fn set_commission_rate<T: Into<u64>>(&mut self, field: T)

Sets commission_rate with the provided value.

pub fn with_commission_rate<T: Into<u64>>(self, field: T) -> Self

Sets commission_rate with the provided value.

pub fn next_epoch_stake_opt_mut(&mut self) -> Option<&mut u64>

If next_epoch_stake is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn next_epoch_stake_mut(&mut self) -> &mut u64

Returns a mutable reference to next_epoch_stake. If the field is unset, it is first initialized with the default value.

pub fn next_epoch_stake_opt(&self) -> Option<u64>

If next_epoch_stake is set, returns Some with the value; otherwise returns None.

pub fn set_next_epoch_stake<T: Into<u64>>(&mut self, field: T)

Sets next_epoch_stake with the provided value.

pub fn with_next_epoch_stake<T: Into<u64>>(self, field: T) -> Self

Sets next_epoch_stake with the provided value.

pub fn next_epoch_gas_price_opt_mut(&mut self) -> Option<&mut u64>

If next_epoch_gas_price is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn next_epoch_gas_price_mut(&mut self) -> &mut u64

Returns a mutable reference to next_epoch_gas_price. If the field is unset, it is first initialized with the default value.

pub fn next_epoch_gas_price_opt(&self) -> Option<u64>

If next_epoch_gas_price is set, returns Some with the value; otherwise returns None.

pub fn set_next_epoch_gas_price<T: Into<u64>>(&mut self, field: T)

Sets next_epoch_gas_price with the provided value.

pub fn with_next_epoch_gas_price<T: Into<u64>>(self, field: T) -> Self

Sets next_epoch_gas_price with the provided value.

pub fn next_epoch_commission_rate_opt_mut(&mut self) -> Option<&mut u64>

If next_epoch_commission_rate is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn next_epoch_commission_rate_mut(&mut self) -> &mut u64

Returns a mutable reference to next_epoch_commission_rate. If the field is unset, it is first initialized with the default value.

pub fn next_epoch_commission_rate_opt(&self) -> Option<u64>

If next_epoch_commission_rate is set, returns Some with the value; otherwise returns None.

pub fn set_next_epoch_commission_rate<T: Into<u64>>(&mut self, field: T)

Sets next_epoch_commission_rate with the provided value.

pub fn with_next_epoch_commission_rate<T: Into<u64>>(self, field: T) -> Self

Sets next_epoch_commission_rate with the provided value.

pub fn extra_fields(&self) -> &MoveTable

Returns the value of extra_fields, or the default value if extra_fields is unset.

pub fn extra_fields_opt_mut(&mut self) -> Option<&mut MoveTable>

If extra_fields is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn extra_fields_mut(&mut self) -> &mut MoveTable

Returns a mutable reference to extra_fields. If the field is unset, it is first initialized with the default value.

pub fn extra_fields_opt(&self) -> Option<&MoveTable>

If extra_fields is set, returns Some with the value; otherwise returns None.

pub fn set_extra_fields<T: Into<MoveTable>>(&mut self, field: T)

Sets extra_fields with the provided value.

pub fn with_extra_fields<T: Into<MoveTable>>(self, field: T) -> Self

Sets extra_fields with the provided value.

impl Validator

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn address(&self) -> &str

Returns the value of address, or the default value if address is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn image_url(&self) -> &str

Returns the value of image_url, or the default value if image_url is unset.

pub fn project_url(&self) -> &str

Returns the value of project_url, or the default value if project_url is unset.

pub fn protocol_public_key(&self) -> &[u8]

Returns the value of protocol_public_key, or the default value if protocol_public_key is unset.

pub fn proof_of_possession(&self) -> &[u8]

Returns the value of proof_of_possession, or the default value if proof_of_possession is unset.

pub fn network_public_key(&self) -> &[u8]

Returns the value of network_public_key, or the default value if network_public_key is unset.

pub fn worker_public_key(&self) -> &[u8]

Returns the value of worker_public_key, or the default value if worker_public_key is unset.

pub fn network_address(&self) -> &str

Returns the value of network_address, or the default value if network_address is unset.

pub fn p2p_address(&self) -> &str

Returns the value of p2p_address, or the default value if p2p_address is unset.

pub fn primary_address(&self) -> &str

Returns the value of primary_address, or the default value if primary_address is unset.

pub fn worker_address(&self) -> &str

Returns the value of worker_address, or the default value if worker_address is unset.

pub fn next_epoch_protocol_public_key(&self) -> &[u8]

Returns the value of next_epoch_protocol_public_key, or the default value if next_epoch_protocol_public_key is unset.

pub fn next_epoch_proof_of_possession(&self) -> &[u8]

Returns the value of next_epoch_proof_of_possession, or the default value if next_epoch_proof_of_possession is unset.

pub fn next_epoch_network_public_key(&self) -> &[u8]

Returns the value of next_epoch_network_public_key, or the default value if next_epoch_network_public_key is unset.

pub fn next_epoch_worker_public_key(&self) -> &[u8]

Returns the value of next_epoch_worker_public_key, or the default value if next_epoch_worker_public_key is unset.

pub fn next_epoch_network_address(&self) -> &str

Returns the value of next_epoch_network_address, or the default value if next_epoch_network_address is unset.

pub fn next_epoch_p2p_address(&self) -> &str

Returns the value of next_epoch_p2p_address, or the default value if next_epoch_p2p_address is unset.

pub fn next_epoch_primary_address(&self) -> &str

Returns the value of next_epoch_primary_address, or the default value if next_epoch_primary_address is unset.

pub fn next_epoch_worker_address(&self) -> &str

Returns the value of next_epoch_worker_address, or the default value if next_epoch_worker_address is unset.

pub fn voting_power(&self) -> u64

Returns the value of voting_power, or the default value if voting_power is unset.

pub fn operation_cap_id(&self) -> &str

Returns the value of operation_cap_id, or the default value if operation_cap_id is unset.

pub fn gas_price(&self) -> u64

Returns the value of gas_price, or the default value if gas_price is unset.

pub fn commission_rate(&self) -> u64

Returns the value of commission_rate, or the default value if commission_rate is unset.

pub fn next_epoch_stake(&self) -> u64

Returns the value of next_epoch_stake, or the default value if next_epoch_stake is unset.

pub fn next_epoch_gas_price(&self) -> u64

Returns the value of next_epoch_gas_price, or the default value if next_epoch_gas_price is unset.

pub fn next_epoch_commission_rate(&self) -> u64

Returns the value of next_epoch_commission_rate, or the default value if next_epoch_commission_rate is unset.