sui_rpc::proto::sui::rpc::v2beta2

Struct SystemState

#[non_exhaustive]
pub struct SystemState {

    pub version: Option<u64>,
    pub epoch: Option<u64>,
    pub protocol_version: Option<u64>,
    pub validators: Option<ValidatorSet>,
    pub storage_fund: Option<StorageFund>,
    pub parameters: Option<SystemParameters>,
    pub reference_gas_price: Option<u64>,
    pub validator_report_records: Vec<ValidatorReportRecord>,
    pub stake_subsidy: Option<StakeSubsidy>,
    pub safe_mode: Option<bool>,
    pub safe_mode_storage_rewards: Option<u64>,
    pub safe_mode_computation_rewards: Option<u64>,
    pub safe_mode_storage_rebates: Option<u64>,
    pub safe_mode_non_refundable_storage_fee: Option<u64>,
    pub epoch_start_timestamp_ms: Option<u64>,
    pub extra_fields: Option<MoveTable>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

version: Option<u64>

The version of the system state data structure type.

epoch: Option<u64>

The epoch id

protocol_version: Option<u64>

The protocol version

validators: Option<ValidatorSet>

Information about the validators

storage_fund: Option<StorageFund>

Storage Fund info

parameters: Option<SystemParameters>

Set of system config parameters

reference_gas_price: Option<u64>

The reference gas price for this epoch

validator_report_records: Vec<ValidatorReportRecord>

A list of the records of validator reporting each other.

There is an entry in this list for each validator that has been reported at least once. Each record contains all the validators that reported them. If a validator has never been reported they don’t have a record in this list. This lists persists across epoch: a peer continues being in a reported state until the reporter doesn’t explicitly remove their report.

stake_subsidy: Option<StakeSubsidy>

Schedule of stake subsidies given out each epoch.

safe_mode: Option<bool>

Whether the system is running in a downgraded safe mode due to a non-recoverable bug. This is set whenever we failed to execute advance_epoch, and ended up executing advance_epoch_safe_mode. It can be reset once we are able to successfully execute advance_epoch. The rest of the fields starting with safe_mode_ are accumulated during safe mode when advance_epoch_safe_mode is executed. They will eventually be processed once we are out of safe mode.

safe_mode_storage_rewards: Option<u64>

Storage rewards accumulated during safe_mode

safe_mode_computation_rewards: Option<u64>

Computation rewards accumulated during safe_mode

safe_mode_storage_rebates: Option<u64>

Storage rebates paid out during safe_mode

safe_mode_non_refundable_storage_fee: Option<u64>

Nonrefundable storage fees accumulated during safe_mode

epoch_start_timestamp_ms: Option<u64>

Unix timestamp of when this this epoch started

extra_fields: Option<MoveTable>

Any extra fields that’s not defined statically.

Implementations

impl SystemState

pub const VERSION_FIELD: &'static MessageField

pub const EPOCH_FIELD: &'static MessageField

pub const PROTOCOL_VERSION_FIELD: &'static MessageField

pub const VALIDATORS_FIELD: &'static MessageField

pub const STORAGE_FUND_FIELD: &'static MessageField

pub const PARAMETERS_FIELD: &'static MessageField

pub const REFERENCE_GAS_PRICE_FIELD: &'static MessageField

pub const VALIDATOR_REPORT_RECORDS_FIELD: &'static MessageField

pub const STAKE_SUBSIDY_FIELD: &'static MessageField

pub const SAFE_MODE_FIELD: &'static MessageField

pub const SAFE_MODE_STORAGE_REWARDS_FIELD: &'static MessageField

pub const SAFE_MODE_COMPUTATION_REWARDS_FIELD: &'static MessageField

pub const SAFE_MODE_STORAGE_REBATES_FIELD: &'static MessageField

pub const SAFE_MODE_NON_REFUNDABLE_STORAGE_FEE_FIELD: &'static MessageField

pub const EPOCH_START_TIMESTAMP_MS_FIELD: &'static MessageField

pub const EXTRA_FIELDS_FIELD: &'static MessageField

impl SystemState

pub fn path_builder() -> SystemStateFieldPathBuilder

impl SystemState

pub const fn const_default() -> Self

pub fn version_opt_mut(&mut self) -> Option<&mut u64>

If version is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn version_mut(&mut self) -> &mut u64

Returns a mutable reference to version. If the field is unset, it is first initialized with the default value.

pub fn version_opt(&self) -> Option<u64>

If version is set, returns Some with the value; otherwise returns None.

pub fn set_version<T: Into<u64>>(&mut self, field: T)

Sets version with the provided value.

pub fn with_version<T: Into<u64>>(self, field: T) -> Self

Sets version with the provided value.

pub fn epoch_opt_mut(&mut self) -> Option<&mut u64>

If epoch is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn epoch_mut(&mut self) -> &mut u64

Returns a mutable reference to epoch. If the field is unset, it is first initialized with the default value.

pub fn epoch_opt(&self) -> Option<u64>

If epoch is set, returns Some with the value; otherwise returns None.

pub fn set_epoch<T: Into<u64>>(&mut self, field: T)

Sets epoch with the provided value.

pub fn with_epoch<T: Into<u64>>(self, field: T) -> Self

Sets epoch with the provided value.

pub fn protocol_version_opt_mut(&mut self) -> Option<&mut u64>

If protocol_version is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn protocol_version_mut(&mut self) -> &mut u64

Returns a mutable reference to protocol_version. If the field is unset, it is first initialized with the default value.

pub fn protocol_version_opt(&self) -> Option<u64>

If protocol_version is set, returns Some with the value; otherwise returns None.

pub fn set_protocol_version<T: Into<u64>>(&mut self, field: T)

Sets protocol_version with the provided value.

pub fn with_protocol_version<T: Into<u64>>(self, field: T) -> Self

Sets protocol_version with the provided value.

pub fn validators(&self) -> &ValidatorSet

Returns the value of validators, or the default value if validators is unset.

pub fn validators_opt_mut(&mut self) -> Option<&mut ValidatorSet>

If validators is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn validators_mut(&mut self) -> &mut ValidatorSet

Returns a mutable reference to validators. If the field is unset, it is first initialized with the default value.

pub fn validators_opt(&self) -> Option<&ValidatorSet>

If validators is set, returns Some with the value; otherwise returns None.

pub fn set_validators<T: Into<ValidatorSet>>(&mut self, field: T)

Sets validators with the provided value.

pub fn with_validators<T: Into<ValidatorSet>>(self, field: T) -> Self

Sets validators with the provided value.

pub fn storage_fund(&self) -> &StorageFund

Returns the value of storage_fund, or the default value if storage_fund is unset.

pub fn storage_fund_opt_mut(&mut self) -> Option<&mut StorageFund>

If storage_fund is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn storage_fund_mut(&mut self) -> &mut StorageFund

Returns a mutable reference to storage_fund. If the field is unset, it is first initialized with the default value.

pub fn storage_fund_opt(&self) -> Option<&StorageFund>

If storage_fund is set, returns Some with the value; otherwise returns None.

pub fn set_storage_fund<T: Into<StorageFund>>(&mut self, field: T)

Sets storage_fund with the provided value.

pub fn with_storage_fund<T: Into<StorageFund>>(self, field: T) -> Self

Sets storage_fund with the provided value.

pub fn parameters(&self) -> &SystemParameters

Returns the value of parameters, or the default value if parameters is unset.

pub fn parameters_opt_mut(&mut self) -> Option<&mut SystemParameters>

If parameters is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn parameters_mut(&mut self) -> &mut SystemParameters

Returns a mutable reference to parameters. If the field is unset, it is first initialized with the default value.

pub fn parameters_opt(&self) -> Option<&SystemParameters>

If parameters is set, returns Some with the value; otherwise returns None.

pub fn set_parameters<T: Into<SystemParameters>>(&mut self, field: T)

Sets parameters with the provided value.

pub fn with_parameters<T: Into<SystemParameters>>(self, field: T) -> Self

Sets parameters with the provided value.

pub fn reference_gas_price_opt_mut(&mut self) -> Option<&mut u64>

If reference_gas_price is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn reference_gas_price_mut(&mut self) -> &mut u64

Returns a mutable reference to reference_gas_price. If the field is unset, it is first initialized with the default value.

pub fn reference_gas_price_opt(&self) -> Option<u64>

If reference_gas_price is set, returns Some with the value; otherwise returns None.

pub fn set_reference_gas_price<T: Into<u64>>(&mut self, field: T)

Sets reference_gas_price with the provided value.

pub fn with_reference_gas_price<T: Into<u64>>(self, field: T) -> Self

Sets reference_gas_price with the provided value.

pub fn validator_report_records(&self) -> &[ValidatorReportRecord]

Returns the value of validator_report_records, or the default value if validator_report_records is unset.

pub fn validator_report_records_mut( &mut self, ) -> &mut Vec<ValidatorReportRecord>

Returns a mutable reference to validator_report_records. If the field is unset, it is first initialized with the default value.

pub fn set_validator_report_records( &mut self, field: Vec<ValidatorReportRecord>, )

Sets validator_report_records with the provided value.

pub fn with_validator_report_records( self, field: Vec<ValidatorReportRecord>, ) -> Self

Sets validator_report_records with the provided value.

pub fn stake_subsidy(&self) -> &StakeSubsidy

Returns the value of stake_subsidy, or the default value if stake_subsidy is unset.

pub fn stake_subsidy_opt_mut(&mut self) -> Option<&mut StakeSubsidy>

If stake_subsidy is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn stake_subsidy_mut(&mut self) -> &mut StakeSubsidy

Returns a mutable reference to stake_subsidy. If the field is unset, it is first initialized with the default value.

pub fn stake_subsidy_opt(&self) -> Option<&StakeSubsidy>

If stake_subsidy is set, returns Some with the value; otherwise returns None.

pub fn set_stake_subsidy<T: Into<StakeSubsidy>>(&mut self, field: T)

Sets stake_subsidy with the provided value.

pub fn with_stake_subsidy<T: Into<StakeSubsidy>>(self, field: T) -> Self

Sets stake_subsidy with the provided value.

pub fn safe_mode_opt_mut(&mut self) -> Option<&mut bool>

If safe_mode is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn safe_mode_mut(&mut self) -> &mut bool

Returns a mutable reference to safe_mode. If the field is unset, it is first initialized with the default value.

pub fn safe_mode_opt(&self) -> Option<bool>

If safe_mode is set, returns Some with the value; otherwise returns None.

pub fn set_safe_mode<T: Into<bool>>(&mut self, field: T)

Sets safe_mode with the provided value.

pub fn with_safe_mode<T: Into<bool>>(self, field: T) -> Self

Sets safe_mode with the provided value.

pub fn safe_mode_storage_rewards_opt_mut(&mut self) -> Option<&mut u64>

If safe_mode_storage_rewards is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn safe_mode_storage_rewards_mut(&mut self) -> &mut u64

Returns a mutable reference to safe_mode_storage_rewards. If the field is unset, it is first initialized with the default value.

pub fn safe_mode_storage_rewards_opt(&self) -> Option<u64>

If safe_mode_storage_rewards is set, returns Some with the value; otherwise returns None.

pub fn set_safe_mode_storage_rewards<T: Into<u64>>(&mut self, field: T)

Sets safe_mode_storage_rewards with the provided value.

pub fn with_safe_mode_storage_rewards<T: Into<u64>>(self, field: T) -> Self

Sets safe_mode_storage_rewards with the provided value.

pub fn safe_mode_computation_rewards_opt_mut(&mut self) -> Option<&mut u64>

If safe_mode_computation_rewards is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn safe_mode_computation_rewards_mut(&mut self) -> &mut u64

Returns a mutable reference to safe_mode_computation_rewards. If the field is unset, it is first initialized with the default value.

pub fn safe_mode_computation_rewards_opt(&self) -> Option<u64>

If safe_mode_computation_rewards is set, returns Some with the value; otherwise returns None.

pub fn set_safe_mode_computation_rewards<T: Into<u64>>(&mut self, field: T)

Sets safe_mode_computation_rewards with the provided value.

pub fn with_safe_mode_computation_rewards<T: Into<u64>>(self, field: T) -> Self

Sets safe_mode_computation_rewards with the provided value.

pub fn safe_mode_storage_rebates_opt_mut(&mut self) -> Option<&mut u64>

If safe_mode_storage_rebates is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn safe_mode_storage_rebates_mut(&mut self) -> &mut u64

Returns a mutable reference to safe_mode_storage_rebates. If the field is unset, it is first initialized with the default value.

pub fn safe_mode_storage_rebates_opt(&self) -> Option<u64>

If safe_mode_storage_rebates is set, returns Some with the value; otherwise returns None.

pub fn set_safe_mode_storage_rebates<T: Into<u64>>(&mut self, field: T)

Sets safe_mode_storage_rebates with the provided value.

pub fn with_safe_mode_storage_rebates<T: Into<u64>>(self, field: T) -> Self

Sets safe_mode_storage_rebates with the provided value.

pub fn safe_mode_non_refundable_storage_fee_opt_mut( &mut self, ) -> Option<&mut u64>

If safe_mode_non_refundable_storage_fee is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn safe_mode_non_refundable_storage_fee_mut(&mut self) -> &mut u64

Returns a mutable reference to safe_mode_non_refundable_storage_fee. If the field is unset, it is first initialized with the default value.

pub fn safe_mode_non_refundable_storage_fee_opt(&self) -> Option<u64>

If safe_mode_non_refundable_storage_fee is set, returns Some with the value; otherwise returns None.

pub fn set_safe_mode_non_refundable_storage_fee<T: Into<u64>>( &mut self, field: T, )

Sets safe_mode_non_refundable_storage_fee with the provided value.

pub fn with_safe_mode_non_refundable_storage_fee<T: Into<u64>>( self, field: T, ) -> Self

Sets safe_mode_non_refundable_storage_fee with the provided value.

pub fn epoch_start_timestamp_ms_opt_mut(&mut self) -> Option<&mut u64>

If epoch_start_timestamp_ms is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn epoch_start_timestamp_ms_mut(&mut self) -> &mut u64

Returns a mutable reference to epoch_start_timestamp_ms. If the field is unset, it is first initialized with the default value.

pub fn epoch_start_timestamp_ms_opt(&self) -> Option<u64>

If epoch_start_timestamp_ms is set, returns Some with the value; otherwise returns None.

pub fn set_epoch_start_timestamp_ms<T: Into<u64>>(&mut self, field: T)

Sets epoch_start_timestamp_ms with the provided value.

pub fn with_epoch_start_timestamp_ms<T: Into<u64>>(self, field: T) -> Self

Sets epoch_start_timestamp_ms with the provided value.

pub fn extra_fields(&self) -> &MoveTable

Returns the value of extra_fields, or the default value if extra_fields is unset.

pub fn extra_fields_opt_mut(&mut self) -> Option<&mut MoveTable>

If extra_fields is set, returns Some with a mutable reference to the value; otherwise returns None.

pub fn extra_fields_mut(&mut self) -> &mut MoveTable

Returns a mutable reference to extra_fields. If the field is unset, it is first initialized with the default value.

pub fn extra_fields_opt(&self) -> Option<&MoveTable>

If extra_fields is set, returns Some with the value; otherwise returns None.

pub fn set_extra_fields<T: Into<MoveTable>>(&mut self, field: T)

Sets extra_fields with the provided value.

pub fn with_extra_fields<T: Into<MoveTable>>(self, field: T) -> Self

Sets extra_fields with the provided value.

impl SystemState

pub fn version(&self) -> u64

Returns the value of version, or the default value if version is unset.

pub fn epoch(&self) -> u64

Returns the value of epoch, or the default value if epoch is unset.

pub fn protocol_version(&self) -> u64

Returns the value of protocol_version, or the default value if protocol_version is unset.

pub fn reference_gas_price(&self) -> u64

Returns the value of reference_gas_price, or the default value if reference_gas_price is unset.

pub fn safe_mode(&self) -> bool

Returns the value of safe_mode, or the default value if safe_mode is unset.

pub fn safe_mode_storage_rewards(&self) -> u64

Returns the value of safe_mode_storage_rewards, or the default value if safe_mode_storage_rewards is unset.

pub fn safe_mode_computation_rewards(&self) -> u64

Returns the value of safe_mode_computation_rewards, or the default value if safe_mode_computation_rewards is unset.

pub fn safe_mode_storage_rebates(&self) -> u64

Returns the value of safe_mode_storage_rebates, or the default value if safe_mode_storage_rebates is unset.

pub fn safe_mode_non_refundable_storage_fee(&self) -> u64

Returns the value of safe_mode_non_refundable_storage_fee, or the default value if safe_mode_non_refundable_storage_fee is unset.

pub fn epoch_start_timestamp_ms(&self) -> u64

Returns the value of epoch_start_timestamp_ms, or the default value if epoch_start_timestamp_ms is unset.